MaintainX/MaintenX/requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
APScheduler==3.10.4
reportlab==4.2.2